Added new factory.
